chore: Update database names in constant.js files
This commit updates the database names in the constant.js files for the 04-youtube-api and 05-ecommerce-api modules. The DB_NAME constant in 04-youtube-api/src/constant.js is changed to "youtube-db", while the DB_NAME constant in 05-ecommerce-api/src/constant.js is changed to "ecommerce-db". This change ensures that the correct database is used for each module.